write tests

+++ contracts/HotelManager.sol
@@ -1,15 +1,15 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity >=0.7.0 <0.9.0;
 
-import "./IResourceManager.sol"; 
+import "./IResourceManager.sol";
 import {StringUtils} from "./StringUtils.sol";
 
 contract HotelManager {
     address private resourceManagerAddress;
 
     event IsRoomAvailable(string txId, bool isRoomAvailable);
-    event QueryRoomPrice(string txId, uint roomPrice);
-    event QueryClientBalance(string txId, uint clientBalance);
+    event QueryRoomPrice(string txId, uint256 roomPrice);
+    event QueryClientBalance(string txId, uint256 clientBalance);
     event HasReservation(string txId, bool hasReservation);
 
     function setResourceManagerAddress(address _newAddress) public {
@@ -28,12 +28,12 @@ contract HotelManager {
         return result;
     }
 
-    function queryRoomPrice(string calldata txId) external returns (uint) {
-        uint defaultPrice = 500;
+    function queryRoomPrice(string calldata txId) external returns (uint256) {
+        uint256 defaultPrice = 500;
         string memory priceS = getRM().getValue(txId, "roomPrice");
-        uint result = defaultPrice;
+        uint256 result = defaultPrice;
 
-        if(!StringUtils.isEmpty(priceS)) {
+        if (!StringUtils.isEmpty(priceS)) {
             result = StringUtils.stringToUint(priceS);
         }
 
@@ -41,10 +41,10 @@ contract HotelManager {
         return result;
     }
 
-    function queryClientBalance(string calldata txId) public returns (uint) {
+    function queryClientBalance(string calldata txId) public returns (uint256) {
         string memory varName = formulateClientBalanceVarName(tx.origin);
         string memory balance = getRM().getValue(txId, varName);
-        uint result = 0;
+        uint256 result = 0;
 
         if (!StringUtils.isEmpty(balance)) {
             result = StringUtils.stringToUint(balance);
@@ -54,25 +54,31 @@ contract HotelManager {
         return result;
     }
 
-    function changeRoomPrice(string calldata txId, uint newPrice) external {
+    function changeRoomPrice(string calldata txId, uint256 newPrice) external {
         string memory priceS = StringUtils.uintToString(newPrice);
         getRM().setValue(txId, "roomPrice", priceS);
     }
 
-    function addToClientBalance(string calldata txId, uint amountToAdd) external {
+    function addToClientBalance(string calldata txId, uint256 amountToAdd)
+        external
+    {
         require(amountToAdd > 0, "The amount must be a positive value!");
         string memory varName = formulateClientBalanceVarName(tx.origin);
-        uint balance = queryClientBalance(txId);
-        uint newBalance = balance + amountToAdd;
+        uint256 balance = queryClientBalance(txId);
+        uint256 newBalance = balance + amountToAdd;
         string memory newBalanceS = StringUtils.uintToString(newBalance);
         getRM().setValue(txId, varName, newBalanceS);
     }
 
     function bookRoom(string calldata txId) external {
         require(isRoomAvailable(txId), "the room must be available!");
-        uint roomPrice = this.queryRoomPrice(txId);
+        uint256 roomPrice = this.queryRoomPrice(txId);
         deductFromClientBalance(txId, roomPrice);
-        getRM().setValue(txId, "roomOwner", StringUtils.addressToHexString(tx.origin));
+        getRM().setValue(
+            txId,
+            "roomOwner",
+            StringUtils.addressToHexString(tx.origin)
+        );
     }
 
     function hasReservation(string calldata txId) external returns (bool) {
@@ -83,17 +89,27 @@ contract HotelManager {
         emit HasReservation(txId, result);
         return result;
     }
- 
-    function deductFromClientBalance(string calldata txId, uint amountToDeduct) internal {
+
+    function deductFromClientBalance(
+        string calldata txId,
+        uint256 amountToDeduct
+    ) internal {
         string memory varName = formulateClientBalanceVarName(tx.origin);
-        uint balance = queryClientBalance(txId);
-        uint newBalance = balance - amountToDeduct;
-        require(newBalance >= 0, "The amount to deduct cannot be larger than the current balance!");
+        uint256 balance = queryClientBalance(txId);
+        uint256 newBalance = balance - amountToDeduct;
+        require(
+            newBalance >= 0,
+            "The amount to deduct cannot be larger than the current balance!"
+        );
         string memory newBalanceS = StringUtils.uintToString(newBalance);
         getRM().setValue(txId, varName, newBalanceS);
     }
 
-    function formulateClientBalanceVarName(address client) private pure returns (string memory) {    
+    function formulateClientBalanceVarName(address client)
+        private
+        pure
+        returns (string memory)
+    {
         return StringUtils.addressToHexString(client);
     }
 }
